Slight fix.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31374 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/ssi2001-drv.c
@@ -112,7 +112,7 @@ static void ssi2001_outb(unsigned int addrint, short value)
     }
 }
 
-static short ssi2001_inb(unsigned int addrint)
+static BYTE ssi2001_inb(unsigned int addrint)
 {
     WORD addr = (WORD)addrint;
     DWORD tmp;
